separated out DisInstM into separate module
